refactor templates
